refactor search to store
